fix: move section1 calcultions to own file
